General clean up and optimization

+++ src/Renderer/index_buffer.cpp
@@ -3,7 +3,7 @@
 using namespace Inferonix::Renderer;
 
 
-index_buffer::index_buffer(size_t count, const void* indices)
+index_buffer::index_buffer(int64_t count, const void* indices)
     : buffer(), _count(count)
 {}
 
@@ -17,7 +17,7 @@ void index_buffer::unbind() const
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 }
 
-void index_buffer::buffer_data(size_t count, const void * indices)
+void index_buffer::buffer_data(int64_t count, const void * indices)
 {
     glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLsizeiptr>(count), indices, GL_STATIC_DRAW);
 }
